requirements.txt
numpy 
pandas
pymongo
pymongo[srv]==3.12
certifi
python-dotenv
dill
scikit-learn
pyaml
mlflow
fastapi 
uvicorn
